requirements.txt
langchain==0.3.27
langgraph==0.6.10
langchain-google-genai==2.1.12
langchain-community==0.3.31
duckduckgo-search==8.1.1
streamlit==1.50.0
python-dotenv==1.1.1
google-search-results==2.4.2
tavily-python==0.7.12

